requirements.txt
adafruit-blinka
adafruit-circuitpython-dht
RPi.GPIO
matplotlib
fastapi
uvicorn
